LPS-179094 Extract ScreenNavigationEntry methods from ObjectDefinitionsActionsScreenNavigationCategory to ObjectDefinitionsActionsScreenNavigationEntry, implement getCategoryKey method in ObjectDefinitionsActionsScreenNavigationEntry and make ObjectDefinitionsActionsScreenNavigationCategory extends BaseObjectDefinitionsScreenNavigationCategory
