+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/definitions/frontend/taglib/servlet/taglib/ObjectDefinitionsActionsScreenNavigationCategory.java
@@ -15,111 +15,24 @@
 package com.liferay.object.web.internal.object.definitions.frontend.taglib.servlet.taglib;
 
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.notification.service.NotificationTemplateLocalService;
-import com.liferay.object.action.executor.ObjectActionExecutorRegistry;
-import com.liferay.object.action.trigger.ObjectActionTriggerRegistry;
-import com.liferay.object.model.ObjectDefinition;
-import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.object.web.internal.object.definitions.constants.ObjectDefinitionsScreenNavigationEntryConstants;
-import com.liferay.object.web.internal.object.definitions.display.context.ObjectDefinitionsActionsDisplayContext;
-import com.liferay.portal.kernel.json.JSONFactory;
-import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
-
-import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Marco Leo
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=50",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=50",
+	service = ScreenNavigationCategory.class
 )
 public class ObjectDefinitionsActionsScreenNavigationCategory
-	extends BaseObjectDefinitionsScreenNavigationEntry
-	implements ScreenNavigationCategory {
+	extends BaseObjectDefinitionsScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
 		return ObjectDefinitionsScreenNavigationEntryConstants.
 			CATEGORY_KEY_ACTIONS;
 	}
-
-	@Override
-	public String getEntryKey() {
-		return ObjectDefinitionsScreenNavigationEntryConstants.
-			ENTRY_KEY_ACTIONS;
-	}
-
-	@Override
-	public String getJspPath() {
-		return "/object_definitions/object_definition/actions.jsp";
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return _language.get(locale, "actions");
-	}
-
-	@Override
-	public boolean isVisible(User user, ObjectDefinition objectDefinition) {
-		return objectDefinition.isDefaultStorageType();
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			new ObjectDefinitionsActionsDisplayContext(
-				httpServletRequest, _jsonFactory,
-				_notificationTemplateLocalService,
-				_objectActionExecutorRegistry, _objectActionTriggerRegistry,
-				_objectDefinitionLocalService,
-				_objectDefinitionModelResourcePermission));
-
-		super.render(httpServletRequest, httpServletResponse);
-	}
-
-	@Reference
-	private JSONFactory _jsonFactory;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private NotificationTemplateLocalService _notificationTemplateLocalService;
-
-	@Reference
-	private ObjectActionExecutorRegistry _objectActionExecutorRegistry;
-
-	@Reference
-	private ObjectActionTriggerRegistry _objectActionTriggerRegistry;
-
-	@Reference
-	private ObjectDefinitionLocalService _objectDefinitionLocalService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.object.model.ObjectDefinition)"
-	)
-	private ModelResourcePermission<ObjectDefinition>
-		_objectDefinitionModelResourcePermission;
 
 }
